MCP-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.182.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.1",
]
